Added the layer class to the core functionality
